feat(slots): extract @derive/data-slots + the read contract (prototype)
Phase 3, prototyped on top of #580 rather than shipped. Nothing here is
published as a standard, and it should not be until somebody outside the
team emits a slot — publishing a spec for a format nothing emits is how
CSVW and W3C Web Annotation died, and both were better specified than the
things that beat them.

Writing the parser is what showed the parser is the least valuable part:
a few hundred lines, no incumbent, nothing an implementer would get wrong.
SPEC.md is the actual contribution — the read contract, where every item
is a decision an independent host must make and the wrong choice fails
silently. It is written from THIS host's live defects: absence is not a
zero, cross-record rows join to the current version, aggregates disclose
so gate before counting, never report what was filtered, a caller-specific
response is not shared-cacheable, and a sandboxed page needs CORS to read
its own data (and still cannot if the record is gated).

Extraction is a MOVE, not a copy: core re-exports the package, so every
consumer is unchanged and there is one implementation. A second copy is
the failure this codebase has now shipped three times.

lint:slots-portable keeps the claim true — zero imports, zero runtime deps
— and fails loudly if it matches no source at all. Verified against a
known-bad import before wiring in, and the package was bundled standalone
(8.8KB, esbuild, bare Node, no repo) and run to prove it is copyable.
