+++ packages/data-slots/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@derive/core",
+  "name": "@derive/data-slots",
   "version": "0.1.0",
   "private": true,
   "type": "module",
@@ -7,7 +7,7 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/derive-to/derive.git",
-    "directory": "packages/core"
+    "directory": "packages/data-slots"
   },
   "exports": {
     ".": "./src/index.ts"
@@ -17,13 +17,6 @@
     "typecheck:tsc": "tsc --noEmit",
     "test": "vitest run --passWithNoTests",
     "test:coverage": "vitest run --coverage --passWithNoTests"
-  },
-  "dependencies": {
-    "@derive/data-slots": "workspace:*",
-    "fflate": "^0.8.2",
-    "marked": "^18.0.5",
-    "nanoid": "^5.0.0",
-    "xss": "^1.0.15"
   },
   "devDependencies": {
     "typescript": "^6.0.3",
